Message test cleanup and hardening
